iOS changes
* Asynchronous working implemented on iOS
* Metadata returned in the ExtractResult on iOS
* Corrected some bugs in the example

+++ android/src/main/kotlin/cl/ceisufro/fluttermbtilesextractor/Executor.kt
@@ -89,7 +89,7 @@ class Executor private constructor(private var contextReference: WeakReference<C
                 format = it.value
         }
         val name = reader.metadata.tilesetName
-        val version = reader.metadata.tilesetVersion.toInt()
+        val version = reader.metadata.tilesetVersion
         val latitudeSW = reader.metadata.tilesetBounds.bottom
         val longitudeSW = reader.metadata.tilesetBounds.left
         val latitudeNE = reader.metadata.tilesetBounds.top
